* commit 'd2176ea3d7dc675bc128a19323001e3945251af5 [formerly e427756f692593726da4a6eef3ec29f9a2d06d7c]': (342 commits)
  [DF-11684] shaders moved on OS X
  [DF-11684] extra shader removed
  DF-185 fix case-sensitive
  [DF-11685] Fixed render pass desc. format
  [DF-11684] materials and shaders copied from RE to QuickEd
  DF-11682 fix assert in Core::ApplyWindowSize
  [DF-11510] remove profiling from UILayoutSystem.
  DF-11518 fix formatting
  [DF-11591] added +- sign check
  [DF-11591] megasolution build fixed
  [DF-185] format tst fix
  [DF-185] fix build
  [DF-11591] code format only
  [DF-11591] reviewfix
  [DF-11263] comment added
  [DF-11263] fixed bug with russian filenames
  DF-11667 switch ndk
  [DF-11591] reviewfix
  DF-11667 switch ndk
  [DF-11475] format only
  ...


Former-commit-id: ed2a658f46ba097f787112c46e448ac5fc661d52

+++ Sources/Internal/Collision/CollisionObject2.cpp
@@ -35,7 +35,7 @@ void CollisionObject2::SetPolygon(Polygon2* _basePolygon)
     polygon = *_basePolygon;
     basePolygon->CalculateCenterPoint(basePolygonCenter);
     circle.center = basePolygonCenter;
-    circle.radius = sqrtf(basePolygon->CalculateSquareRadius(basePolygonCenter));
+    circle.radius = std::sqrt(basePolygon->CalculateSquareRadius(basePolygonCenter));
 
     forceUpdate = true;
 }
@@ -58,8 +58,8 @@ void CollisionObject2::Update(const Sprite::DrawState & state/*const Vector2 & _
     
     if (type == TYPE_POLYGON)
     {
-        float32 sinA = sinf(angle);
-        float32 cosA = cosf(angle);
+        float32 sinA = std::sin(angle);
+        float32 cosA = std::cos(angle);
         for (int k = 0; k < basePolygon->pointCount; ++k)
         {
             Vector2 * v = &polygon.points[k];
@@ -139,8 +139,8 @@ void CollisionObject2::Update(const Sprite::DrawState& state /*const Vector2 & _
 
     bbox.Empty();
 
-    float32 sinA = sinf(angle);
-    float32 cosA = cosf(angle);
+    float32 sinA = std::sin(angle);
+    float32 cosA = std::cos(angle);
 
     if (type == TYPE_POLYGON)
     {
@@ -159,7 +159,7 @@ void CollisionObject2::Update(const Sprite::DrawState& state /*const Vector2 & _
 
         Vector2 c;
         polygon.CalculateCenterPoint(c);
-        circle.radius = sqrtf(polygon.CalculateSquareRadius(c));
+        circle.radius = std::sqrt(polygon.CalculateSquareRadius(c));
         circle.center = c;
     }
     else
